assessment_proj/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
Django==4.2.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
email_validator==2.2.0
idna==3.7
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pymongo==3.12.3
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
